% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaf.R
\name{scaf}
\alias{scaf}
\title{A Quick Project Creation Function}
\usage{
scaf(project_name)
}
\arguments{
\item{project_name}{A name for your project.}
}
\description{
This function allows you to rapidly create a directory structure and some essential files for an R project.
}
\examples{
scaf("my_project")
}
\keyword{scafford}
